python-modernize

+++ src/ZServer/FTPRequest.py
@@ -113,7 +113,7 @@ class FTPRequest(HTTPRequest):
         elif command == 'STOR':
             env['PATH_INFO'] = self._join_paths(channel.path, path)
             env['REQUEST_METHOD'] = 'PUT'
-            env['CONTENT_LENGTH'] = long(size)  # NOQA
+            env['CONTENT_LENGTH'] = int(size)  # NOQA
         else:
             env['PATH_INFO'] = self._join_paths(channel.path, path, command)
